refactor: Organize DTOs into appropriate packages for better structure
