+++ Backend/src/main/java/com/pg/mbti/dto/auth/LoginResponseDto.java
@@ -1,4 +1,4 @@
-package com.pg.mbti.dto;
+package com.pg.mbti.dto.auth;
 
 public record LoginResponseDto(
         String response
