+++ Backend/src/main/java/com/pg/mbti/dto/auth/RegistrationResponseDto.java
@@ -1,4 +1,4 @@
-package com.pg.mbti.dto;
+package com.pg.mbti.dto.auth;
 
 public record RegistrationResponseDto(
         String username,
