+++ Backend/src/main/java/com/pg/mbti/dto/password/ForgotPasswordDto.java
@@ -1,4 +1,4 @@
-package com.pg.mbti.dto;
+package com.pg.mbti.dto.password;
 
 public record ForgotPasswordDto(
         String email
