+++ Backend/src/main/java/com/pg/mbti/dto/password/ResetPasswordDto.java
@@ -1,4 +1,4 @@
-package com.pg.mbti.dto;
+package com.pg.mbti.dto.password;
 
 public record ResetPasswordDto(
         String token,
