+++ Backend/src/main/java/com/pg/mbti/dto/password/UpdatePasswordDto.java
@@ -1,4 +1,4 @@
-package com.pg.mbti.dto;
+package com.pg.mbti.dto.password;
 
 import lombok.Builder;
